web_app/requirements.txt
pymysql
Flask
pyodbc
Werkzeug
numpy
Pillow
requests
cryptography

